refactor: AI 추천 역주변, 테마선택 영역 컴포넌트화
